context-menu-bin: Fix non-working actions
Destroying the popover on close results in the action not being
called.

Instead do like the members list and keep a single popover for the
whole list. To do that we need to pass a weak reference to the
closest common parent of the list items, via the list view factory.