room-history: Make text reactions smaller data/resources/style.css: Added class for font on applying reaction src/session/content/room_history/message_row/reaction.rs: Added check if reaction contains text or not src/utils.rs: Added regex function for general purpose
Even if it's their primary use in clients is to send emoji reactions, reactions can actually contain any text.
We have to make sure that we ellipsize long texts, and we need to fix the font as it's currently too big.

To fix this issue, I used regex function to check whether reaction contains text or not.

https://gitlab.gnome.org/GNOME/fractal/-/issues/1044

Part-of: <https://gitlab.gnome.org/GNOME/fractal/-/merge_requests/1134>