room-history: Move item activation code to MessageRow
The previous way of tracking message activation was via `GtkListView`'s
"activate" signal, which was not ideal since it required a complicated
management of each items' "activatable" property. This new
implementation is simpler since it just adds a `GtkClickGesture` to the
message rows that require activation and it handles the activation in
the `MessageRow`. This is also useful for implementing a media viewer
opening animation (which is planned in the next commits) because we can
retrieve the actual widget of the media widget being pressed, which was
not possible previously.