room-history: Replace pill by display name in reply quote
The pill in the reply quote attribution (related content sender) turns
out to be confusing as it looks like a mention. Use the same style as
for the display name in a content message row. The fact that it is in a
quoted block makes it less pronounced and thus more clear.

Also use a binding so the user's display name updates if the user
changes it.