room-history: Remove supposed guard against loading
The purpose seems to be to guard when calling load because the timeline
is ready, but the items are not yet loaded in the history.
However that also prevents to load more items when the view is not full.
It is more of a bug to not be able to load events when the view is not
full, than to make one extra request.