user: Do not hold a GWeakRef to Session
GObject uses a singly-linked list to keep track of GWeakRefs to an
object.
When a GWeakRef is dropped, the list needs to be iterated to find the
right one.
In a room with 10k members, it means a list of at least 10k GWeakRefs,
so dropping them all at once takes too much time.

We don't have this issue with strong references.